tests/pyproject.toml
[tool.poetry]
name = "api_projector"
version = "0.1.0"
description = ""
authors = ["Junte Ltd"]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "*"
psycopg2-binary = "*"
python-memcached = "*"
django = "3.*"
django-health-check = "*"
django-cors-headers = "*"
django_extensions = "*"
django-split-settings = "*"
django-model-utils = "*"
django-minio-storage = "*"
django-filter = "*"
djangorestframework = "*"
graphene-django = "*"
graphene-file-upload = "*"
drf-yasg = "*"
jnt-django-admin-tools = "*"
jnt-django-toolbox = "*"
jnt-django-graphene-toolbox = "*"
uWSGI = "*"
sentry-sdk = "*"
injector = "*"
ibmcloudant = "*"
redis = "*"
flower = "*"
fluent-logger = "*"
celery = "4.*"
pillow = "*"
types-PyYAML = "*"
social-auth-core = "^4.1.0"
social-auth-app-django = "^5.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
mypy = "*"
polint = "*"
dennis = "*"
xenon = "*"
safety = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-xdist = "*"
httpretty = "*"
factory-boy = "*"
flake8 = "*"
flake8-pytest = "*"
flake8-django = "*"
flake8-pytest-style = "*"
wemake-python-styleguide = "*"
black = "*"
flake8-class-attributes-order = "*"
types-requests = "*"
types-python-dateutil = "*"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | .*\/migrations
  )/
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
